Separate service.rs to crab.rs + darwinia.rs
